[DEPLOY] 배포를 위한 지도페이지 컴포넌트 분할 (#133)
* [deploy/#126]: 배포를 위한 page.js 코드 수정

* [deploy/#126]: 렌더링을 위한 analysis 컴포넌트 분리

* [deploy/#126]:서버 컴포넌트에서 처리하기 위한 지도 페이지 컴포넌트 분리
