+++ woori-zip/src/app/map/page.js
@@ -1,165 +1,11 @@
 "use client";
-
-import React, { useState, useEffect } from "react";
-import Sidebar from "../../components/domains/map/Sidebar";
-import PropertyList from "../../components/domains/map/PropertyList";
-import MapView from "../../components/domains/map/MapView";
-import PropertyDetails from "../../components/domains/map/PropertyDetails";
-import styles from "../../components/domains/map/map.module.css";
-import NavBar from "../../components/domains/map/NavBar";
-import CategoryMenu from "../../components/domains/map/CategoryMenu";
-import MobileHeader from "../../components/domains/map/MobileHeader";
-import { fetchHouseList, fetchHouseDetails } from "@/app/api/map/houseApi";
-import { useSearchParams } from "next/navigation";
+import React, { Suspense } from "react";
+import MapPageContent from "@/components/domains/map/MapPageContent";
 
 export default function MapPage() {
-  // 파라미터에서 초기값 가져오기
-  const searchParams = useSearchParams();
-  const initialFilters = {
-    level: 7,
-    southWestLatitude: parseFloat(searchParams.get("southWestLatitude")) || 37.5189,
-    southWestLongitude: parseFloat(searchParams.get("southWestLongitude")) || 126.8952,
-    northEastLatitude: parseFloat(searchParams.get("northEastLatitude")) || 37.6157,
-    northEastLongitude: parseFloat(searchParams.get("northEastLongitude")) || 127.0601,
-    category: searchParams.get("category") || null,
-    minDeposit: 0,
-    housetype: "아파트",
-    maxDeposit: 1000000000,
-    minMonthlyRentFee: 0,
-    maxMonthlyRentFee: 2000000000,
-    minMaintenanceFee: 0,
-    maxMaintenanceFee: 5000000,
-  };
-
-  const [mapState, setMapState] = useState(initialFilters);
-  const [houseType, setHouseType] = useState(initialFilters.housetype);
-  const [houseInfo, setHouseInfo] = useState({ houseContents: [], counts: [] });
-  const [houseData, setHouseData] = useState([]);
-  const [mapLocations, setMapLocations] = useState([]);
-  const [selectedProperty, setSelectedProperty] = useState(null);
-  const [isMobile, setIsMobile] = useState(false);
-  const [isSidebarVisible, setSidebarVisible] = useState(false);
-  const [isCategoryMenuVisible, setCategoryMenuVisible] = useState(false);
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => setIsMobile(window.innerWidth <= 393);
-    handleResize();
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
-
-  // 데이터 로드
-  useEffect(() => {
-    const loadInitialData = async () => {
-      try {
-        const data = await fetchHouseList(mapState);
-        console.log("House List Loaded: ", data);
-        updateHouseData(data);
-      } catch (error) {
-        console.error("Failed to load house list: ", error);
-      }
-    };
-    loadInitialData();
-  }, [mapState]);
-
-  const updateHouseData = (data) => {
-    if (data?.houseContents) {
-      setHouseInfo(data);
-      setHouseData(data.houseContents);
-      setMapLocations(
-        data.houseContents.map((house) => ({
-          lat: house.latitude,
-          lng: house.longitude,
-          houseId: house.houseId,
-        }))
-      );
-    }
-  };
-
-  const toggleSidebar = () => setSidebarVisible(!isSidebarVisible);
-  const toggleCategoryMenu = () => setCategoryMenuVisible(!isCategoryMenuVisible);
-  const togglePropertyList = () => setIsExpanded(!isExpanded);
-
-  const handlePropertyClick = async (propertyId) => {
-    try {
-      const data = await fetchHouseDetails(propertyId);
-      setSelectedProperty(data);
-    } catch (error) {
-      console.error("Failed to load property details:", error);
-    }
-  };
-
   return (
-    <div className={styles.container}>
-      <Sidebar houseType={houseType} onSelectCategory={setHouseType} />
-      <div className={styles.mainContent}>
-        <NavBar
-          houseType={houseType}
-          mapState={mapState}
-          onHouseInfoUpdate={updateHouseData}
-        />
-        <div className={styles.contentArea}>
-          {isMobile ? (
-            <MobilePropertyList
-              data={houseData}
-              isExpanded={isExpanded}
-              togglePropertyList={togglePropertyList}
-              onPropertyClick={handlePropertyClick}
-            />
-          ) : (
-            <PropertyList
-              data={houseData}
-              className={styles.webPropertyList}
-              onPropertyClick={handlePropertyClick}
-            />
-          )}
-          <MapView
-            property={selectedProperty}
-            locations={mapLocations}
-            selectedLocation={{ lat: mapState.southWestLatitude, lng: mapState.southWestLongitude }}
-            onMapChange={setMapState}
-            mapViewData={houseInfo}
-          />
-          {selectedProperty && (
-            <PropertyDetails
-              property={selectedProperty}
-              onClose={() => setSelectedProperty(null)}
-            />
-          )}
-        </div>
-      </div>
-      {isCategoryMenuVisible && (
-        <CategoryMenu
-          isVisible={isCategoryMenuVisible}
-          onClose={toggleCategoryMenu}
-          selectedCategory={mapState.category}
-          onApply={(filterData) => {
-            console.log("Applied Filter Data: ", filterData);
-            setMapState((prev) => ({ ...prev, ...filterData }));
-          }}
-        />
-      )}
-      {isMobile && (
-        <MobileHeader
-          onHamburgerClick={toggleSidebar}
-          onFilterClick={toggleCategoryMenu}
-        />
-      )}
-    </div>
-  );
-}
-
-function MobilePropertyList({ data, isExpanded, togglePropertyList, onPropertyClick }) {
-  return (
-    <div
-      className={styles.mobilePropertyList}
-      style={{ height: isExpanded ? "50vh" : "10vh" }}
-    >
-      <div className={styles.handle} onClick={togglePropertyList}>
-        <div className={styles.handleText}>집 목록 보기</div>
-      </div>
-      <PropertyList data={data} onPropertyClick={onPropertyClick} />
-    </div>
+    <Suspense fallback={<div>Loading...</div>}>
+      <MapPageContent />
+    </Suspense>
   );
 }
